\docType{data}
\name{chicagonet}
\alias{chicagonet}
\title{Roads of Chicago (Illinois, USA) close to the University of Chicago}
\format{
A linear network of class \code{linnet}
}
\usage{
data(chicagonet)
}
\description{
A linear network of class \code{linnet} of the roads of Chicago (Illinois, USA) close to the University of Chicago.
It represents the linear network of the Chicago dataset published and analysed in Ang,
Baddeley and Nair (2012). Note that the network adjacency
matrix is stored as a sparse matrix.
}
\examples{

data(chicagonet)

}
\references{
Ang, Q.W., Baddeley, A. and Nair, G. (2012) Geometrically corrected second-order analysis of events on a linear network, with applications to ecology and criminology. Scandinavian Journal of Statistics 39, 591–617.
}
\author{
Nicoletta D'Angelo
}
\keyword{datasets}
