% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.sepstlppm.R
\name{print.sepstlppm}
\alias{print.sepstlppm}
\title{Print of a fitted separable spatio-temporal Poisson process model on a linear network}
\usage{
\method{print}{sepstlppm}(x, ...)
}
\arguments{
\item{x}{An object of class \code{sepstlppm }}

\item{...}{additional unused argument}
}
\description{
The function prints the main information of the fitted model.
}
\examples{
crimesub <- stpm(valenciacrimes$df[101:200, ],
           names = colnames(valenciacrimes$df)[-c(1:3)],
L = valencianet)

mod1 <- sepstlppm(crimesub, spaceformula = ~x ,
                  timeformula = ~ day)
                  
mod1




}
\seealso{
\link{sepstlppm}
}
\author{
Nicoletta D'Angelo
}
