% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.sepstppm.R
\name{print.sepstppm}
\alias{print.sepstppm}
\title{Print of a fitted separable spatio-temporal Poisson process model}
\usage{
\method{print}{sepstppm}(x, ...)
}
\arguments{
\item{x}{An object of class \code{sepstppm}}

\item{...}{additional unused argument}
}
\description{
The function prints the main information of the fitted model.
}
\examples{
crimesub <- stpm(valenciacrimes$df[101:200, ],
           names = colnames(valenciacrimes$df)[-c(1:3)])

mod1 <- sepstppm(crimesub, spaceformula = ~x ,
                  timeformula = ~ day)
mod1




}
\seealso{
\link{sepstppm}
}
\author{
Nicoletta D'Angelo
}
