% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.stlpm.R
\name{print.stlpm}
\alias{print.stlpm}
\title{Print a \code{stlpm} object}
\usage{
\method{print}{stlpm}(x, ...)
}
\arguments{
\item{x}{An object of class \code{stlpm}}

\item{...}{additional unused argument}
}
\description{
It prints the main information on the spatio-temporal point pattern
stored in the \code{stlpm} object: the number of points; the enclosing spatial window;
the temporal time period; information on marks.
}
\examples{

\dontrun{
set.seed(2)
df <- data.frame(x = runif(100, 0, 0.8), y = runif(100, 0, 0.8), t = runif(100), m = rpois(100, 15))

stlpm1 <- stpm(df, L = chicagonet)

stlpm1

}


}
\author{
Nicoletta D'Angelo
}
