% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.stpm.R
\name{print.stpm}
\alias{print.stpm}
\title{Print a \code{stpm} object}
\usage{
\method{print}{stpm}(x, ...)
}
\arguments{
\item{x}{An object of class \code{stpm}}

\item{...}{additional unused argument}
}
\description{
It prints the main information on the spatio-temporal point pattern
stored in the \code{stpm} object: the number of points; the enclosing spatial window;
the temporal time period; information on marks.
}
\examples{

\dontrun{
set.seed(2)
df <- data.frame(cbind(runif(100), runif(100), runif(100), rpois(100, 15),
rpois(100, 30)))

stpm1 <- stpm(df)

summary(stpm1)
}


}
\author{
Nicoletta D'Angelo
}
