% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stpm.R
\name{stpm}
\alias{stpm}
\title{Create \code{stpm} and \code{stlpm} objects for marked point patterns storage}
\usage{
stpm(df, names = NULL, L)
}
\arguments{
\item{df}{A matrix with three columns + m marks}

\item{names}{Factor string to name the marks columns.}

\item{L}{Optional. The linear network of class \code{linnet}.
If provided, the function returns a \code{stlpm} object.}
}
\value{
An \code{stpm} or \code{stlppm} object, depending on whether or not an object of class
linnet is provided for the \code{L} argument.
}
\description{
This function creates a \code{stpm} object as a dataframe with 3 + m columns:
\code{x}, \code{y}, \code{t}, and \code{m} columns to store different marks.
If also the linear network \code{L}, of class linnet, is provided, a \code{stlp}
object is created instead.
}
\examples{


set.seed(2)
df <- data.frame(cbind(runif(100), runif(100), runif(100), rpois(100, 15),
rpois(100, 30)))

stpm1 <- stpm(df)

## Categorical marks

set.seed(2)
dfA <- data.frame(x = runif(100), y = runif(100), t = runif(100), 
                  m1 = rnorm(100), m2 = rep(c("C"), times = 100))
dfB <- data.frame(x = runif(50), y = runif(50), t = runif(50), 
                  m1 = rnorm(25), m2 = rep(c("D"), times = 50))

stpm2 <- stpm(rbind(dfA, dfB), names = c("continuous", "dichotomous"))

## Linear network

set.seed(2)
dfL <- data.frame(cbind(runif(100, 0, 0.85), runif(100, 0, 0.85), runif(100), 
                       rpois(100, 15)))

stlpm1 <- stpm(dfL, L = chicagonet)



}
\author{
Nicoletta D'Angelo
}
