% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.sepstlppm.R
\name{summary.sepstlppm}
\alias{summary.sepstlppm}
\title{Summary of a fitted fitted separable spatio-temporal Poisson process model on a linear network}
\usage{
\method{summary}{sepstlppm}(object, ...)
}
\arguments{
\item{object}{An object of class \code{sepstlppm}}

\item{...}{additional unused argument}
}
\description{
The function summarises the main information of the fitted model.
}
\examples{
\dontrun{
crimesub <- stpm(valenciacrimes$df[101:200, ],
           names = colnames(valenciacrimes$df)[-c(1:3)],
L = valencianet)

mod1 <- sepstlppm(crimesub, spaceformula = ~x ,
                  timeformula = ~ day)
                  
summary(mod1)
}



}
\seealso{
\link{sepstlppm}
}
\author{
Nicoletta D'Angelo
}
