% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.sepstppm.R
\name{summary.sepstppm}
\alias{summary.sepstppm}
\title{Summary of a fitted separable spatio-temporal Poisson process model}
\usage{
\method{summary}{sepstppm}(object, ...)
}
\arguments{
\item{object}{An object of class \code{sepstppm}}

\item{...}{additional unused argument}
}
\description{
The function summarises the main information of the fitted model.
}
\examples{

crimesub <- stpm(valenciacrimes$df[101:200, ],
           names = colnames(valenciacrimes$df)[-c(1:3)])

mod1 <- sepstppm(crimesub, spaceformula = ~x ,
                  timeformula = ~ day)
                  
summary(mod1)



}
\seealso{
\link{sepstlppm}
}
\author{
Nicoletta D'Angelo
}
