\docType{data}
\name{valenciacrimes}
\alias{valenciacrimes}
\title{Crimes in Valencia in 2019}
\format{
A \code{data.frame}
}
\usage{
data(valenciacrimes)
}
\description{
A dataset in \code{data.frame} format containing the 10929 crimes occurred in Valencia, Spain, in 2019.
}
\details{
The variables are as follows:
\itemize{
  \item crime_id.
  \item crime_date.
  \item crime_time.
  \item month.
  \item week.
  \item day.
  \item week_day.
  \item week_day_name.
  \item crime_hour.
  \item crime_lon.
  \item crime_lat.
  \item atm_dist.
  \item bank_dist.
  \item bar_dist.
  \item cafe_dist.
  \item industrial_dist.
  \item market_dist.
  \item nightclub_dist.
  \item police_dist.
  \item pub_dist.
  \item restaurant_dist.
  \item taxi_dist.
  \item x.
  \item y.
  \item t.
}
}
\examples{

data(valenciacrimes)

}
\author{
Nicoletta D'Angelo
}
\keyword{datasets}
