\docType{data}
\name{valencianet}
\alias{valencianet}
\title{Roads of Valencia, Spain}
\format{
A linear network of class \code{linnet}
}
\usage{
data(valencianet)
}
\description{
A linear network of class \code{linnet} of the roads of Valencia, Spain
}
\examples{

data(valencianet)

}
\author{
Nicoletta D'Angelo
}
\keyword{datasets}
