\docType{data}
\name{chicagonet}
\alias{chicagonet}
\title{Rescaled roads of Chicago (Illinois, USA)}
\format{
A linear network of class \code{linnet}
}
\usage{
data(chicagonet)
}
\description{
A linear network of class \code{linnet} of the roads of Chicago (Illinois, USA) close to the University of Chicago.
The window has been rescaled to be enclosed in a unit square.
}
\examples{

data(chicagonet)

}
\references{
Ang, Q.W., Baddeley, A. and Nair, G. (2012) Geometrically corrected second-order analysis of events on a linear network, with applications to ecology and criminology. Scandinavian Journal of Statistics 39, 591–617.
}
\author{
Nicoletta D'Angelo
}
\keyword{datasets}
