% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.stp.R
\name{summary.stp}
\alias{summary.stp}
\title{Summary of a \code{stp} object}
\usage{
\method{summary}{stp}(object, ...)
}
\arguments{
\item{object}{An object of class \code{stp}}

\item{...}{additional unused argument}
}
\description{
It prints the summary statistics of the spatial and temporal coordinates
of  the spatio-temporal point pattern
stored in the \code{stp} object.
}
\examples{


set.seed(2)
df <- data.frame(cbind(runif(100), runif(100), runif(100)))

stp1 <- stp(df)
summary(stp1)



}
\seealso{
\link{stp}, \link{print.stp}, \link{plot.stp}
}
\author{
Nicoletta D'Angelo
}
