\docType{data}
\name{valenciacrimes}
\alias{valenciacrimes}
\title{Crimes in Valencia in 2019}
\format{
A \code{stpm} object
}
\usage{
data(valenciacrimes)
}
\description{
A dataset in \code{stpm} format containing the 10929 crimes occurred in Valencia, Spain, in 2019.
}
\details{
The 15 available marks are the following:
\itemize{
  \item month.
  \item week.
  \item day.
  \item week_day.
  \item atm_dist.
  \item bank_dist.
  \item bar_dist.
  \item cafe_dist.
  \item industrial_dist.
  \item market_dist.
  \item nightclub_dist.
  \item police_dist.
  \item pub_dist.
  \item restaurant_dist.
  \item taxi_dist.
}
}
\examples{

data(valenciacrimes)

}
\author{
Nicoletta D'Angelo
}
\keyword{datasets}
